Python_client/requirements.txt
httpx==0.27.0
pydantic==2.7.1
requests==2.31.0
PyJWT==2.8.0
python-dotenv==1.0.1
